In CoC nvim plugin disable Prettier if not explicitly configured
Prettier is some bullshit JavaScript formatter from Facebook - so it's
probably a malware. I don't like it and would rather not use it at all,
but some clients insist on using it in their projects.

This way at least Prettier won't be executed unless a project has
explicit configuration for it.